refactor(model): split backend.py into base/torch/onnx modules (refs #335)
