--- src/raitap/models/backend.py
@@ -1,428 +1,0 @@
-from __future__ import annotations
-
-import logging
-from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Any, ClassVar, overload
-
-import numpy as np
-
-from raitap.models.registration import register
-from raitap.types import FORWARD_ONLY, Capability, ResolvedHardware, TaskKind
-from raitap.utils.errors import ModelInputShapeError
-from raitap.utils.lazy import lazy_import
-
-from .runtime import _ONNX_RUNTIME_INSTALL_HINT, resolve_onnx_providers
-
-if TYPE_CHECKING:
-    from collections.abc import Callable, Sequence
-    from pathlib import Path
-
-    import onnxruntime as ort
-    import torch
-    from torch import nn
-else:
-    torch = lazy_import("torch")
-    nn = lazy_import("torch.nn")
-
-logger = logging.getLogger(__name__)
-
-
-@overload
-def _adapt_input_shape(
-    inputs: torch.Tensor,
-    expected: tuple[int | None, ...] | None,
-) -> torch.Tensor: ...
-
-
-@overload
-def _adapt_input_shape(
-    inputs: np.ndarray[Any, Any],
-    expected: tuple[int | None, ...] | None,
-) -> np.ndarray[Any, Any]: ...
-
-
-def _adapt_input_shape(
-    inputs: torch.Tensor | np.ndarray[Any, Any],
-    expected: tuple[int | None, ...] | None,
-) -> torch.Tensor | np.ndarray[Any, Any]:
-    """Reshape ``inputs`` to match ``expected`` (with ``None`` = batch dim).
-
-    Returns inputs unchanged when ``expected`` is ``None`` or already matches.
-    Raises :class:`ModelInputShapeError` when per-sample numel mismatches.
-    """
-    if expected is None:
-        return inputs
-
-    input_shape = tuple(int(dim) for dim in inputs.shape)
-    batch = input_shape[0] if input_shape else 1
-    target = tuple(batch if dim is None else int(dim) for dim in expected)
-
-    if input_shape == target:
-        return inputs
-
-    input_numel = 1
-    for dim in input_shape:
-        input_numel *= dim
-    target_numel = 1
-    for dim in target:
-        target_numel *= dim
-
-    if input_numel != target_numel:
-        raise ModelInputShapeError(expected_shape=expected, input_shape=input_shape)
-
-    return inputs.reshape(target)
-
-
-def _resolve_onnx_expected_shape(
-    raw_shape: Sequence[Any],
-) -> tuple[int | None, ...] | None:
-    """Translate an ONNX input shape spec into a ``(int | None, ...)`` tuple
-    suitable for :func:`_adapt_input_shape`.
-
-    ONNX dims can be ints, ``None``, or strings (symbolic names like
-    ``"batch"``). Concrete positive ints are respected; ``None`` and
-    strings become ``None`` (dynamic, resolved at runtime from the input
-    batch dim).
-
-    Returns ``None`` when the shape is empty (scalar input — no adaptation
-    possible). Raises :class:`ModelInputShapeError` when two or more dims
-    are dynamic, since reshape targets become ambiguous and the user must
-    override via ``data.input_metadata.shape``.
-    """
-    if not raw_shape:
-        return None
-    parsed: list[int | None] = []
-    for dim in raw_shape:
-        if isinstance(dim, int) and dim > 0:
-            parsed.append(int(dim))
-        else:
-            parsed.append(None)
-    dynamic_total = sum(1 for dim in parsed if dim is None)
-    if dynamic_total >= 2:
-        raise ModelInputShapeError(expected_shape=tuple(parsed))
-    return tuple(parsed)
-
-
-def _is_torchvision_detection_model(model: nn.Module) -> bool:
-    """Return True for torchvision detection models (Faster R-CNN, RetinaNet, SSD)."""
-    try:
-        from torchvision.models.detection.generalized_rcnn import GeneralizedRCNN
-        from torchvision.models.detection.retinanet import RetinaNet
-        from torchvision.models.detection.ssd import SSD
-    except ImportError:
-        return False
-    return isinstance(model, GeneralizedRCNN | RetinaNet | SSD)
-
-
-_NUMPY_DTYPES_BY_ONNX_TYPE: dict[str, np.dtype[Any]] = {
-    "tensor(float)": np.dtype(np.float32),
-    "tensor(double)": np.dtype(np.float64),
-    "tensor(float16)": np.dtype(np.float16),
-    "tensor(int64)": np.dtype(np.int64),
-    "tensor(int32)": np.dtype(np.int32),
-    "tensor(int16)": np.dtype(np.int16),
-    "tensor(int8)": np.dtype(np.int8),
-    "tensor(uint8)": np.dtype(np.uint8),
-    "tensor(bool)": np.dtype(np.bool_),
-}
-
-
-class ModelBackend(ABC):
-    """Backend-agnostic model runtime interface."""
-
-    provides: ClassVar[frozenset[Capability]] = frozenset()
-    extensions: ClassVar[frozenset[str]] = frozenset()
-    # uv extra installing this backend's runtime, and the hardware values it
-    # ships per-wheel. Set by ``@register``; read import-free by the deps
-    # scanner. ``None``/empty for non-file or single-wheel backends.
-    extra: ClassVar[str | None] = None
-    supported_hardware: ClassVar[frozenset[ResolvedHardware]] = frozenset()
-    # Declared per-sample input shape with ``None`` marking dynamic dims
-    # (typically the batch dim). ``None`` overall = no rank adaptation.
-    expected_input_shape: tuple[int | None, ...] | None = None
-    # Optional id->name table for the model's output classes (e.g. torchvision
-    # detection ``weights.meta["categories"]``). ``None`` when unavailable.
-    category_names: list[str] | None = None
-
-    @classmethod
-    def from_path(
-        cls, path: Path, *, model_cfg: Any, hardware: str, allow_unsafe_pickle: bool = False
-    ) -> ModelBackend:
-        """Construct this backend from a model file.
-
-        Default raises: file loading is not universal (in-memory/test backends
-        never load from a path). File-backed backends declared with
-        ``extensions=`` in ``@register`` override this; the registry calls it
-        from ``model._load_from_path``.
-        """
-        raise NotImplementedError(f"{cls.__name__} does not support construction from a file path.")
-
-    @property
-    @abstractmethod
-    def hardware_label(self) -> str:
-        """Human-readable label for the resolved runtime backend."""
-
-    @property
-    def task_kind(self) -> TaskKind:
-        """Task family this backend serves. Defaults to ``classification``."""
-        return TaskKind.classification
-
-    def _prepare_inputs(self, inputs: torch.Tensor) -> torch.Tensor:
-        """Adapt runtime inputs to this backend's preferred device/layout."""
-        return _adapt_input_shape(inputs, self.expected_input_shape)
-
-    def _prepare_kwargs(self, kwargs: dict[str, Any]) -> dict[str, Any]:
-        """Adapt explainer/runtime kwargs for this backend."""
-        return kwargs
-
-    @abstractmethod
-    def __call__(self, inputs: torch.Tensor) -> Any:
-        """Run inference for ``inputs``."""
-
-    def predict_callable(self) -> Callable[[torch.Tensor], torch.Tensor]:
-        """Forward-only predict fn. Universal shape; defaults to ``__call__``."""
-        return self.__call__
-
-
-@register(
-    provides={Capability.AUTOGRAD},
-    extensions={".pth", ".pt"},
-    extra="torch",
-    supported_hardware={ResolvedHardware.cpu, ResolvedHardware.cuda, ResolvedHardware.xpu},
-)
-class TorchBackend(ModelBackend):
-    """PyTorch-backed model runtime."""
-
-    def __init__(
-        self,
-        model: nn.Module,
-        *,
-        device: torch.device | None = None,
-        task_kind: TaskKind | None = None,
-        category_names: list[str] | None = None,
-    ) -> None:
-        self.model = model
-        self.device = torch.device("cpu") if device is None else device
-        self._task_kind = task_kind if task_kind is not None else self._infer_task_kind(model)
-        self.category_names = category_names
-
-    @classmethod
-    def from_path(
-        cls, path: Path, *, model_cfg: Any, hardware: str, allow_unsafe_pickle: bool = False
-    ) -> ModelBackend:
-        from raitap.models.model import _load_torch_module_from_path  # deferred: import cycle
-        from raitap.models.runtime import resolve_torch_device
-
-        device = resolve_torch_device(hardware)
-        module = _load_torch_module_from_path(
-            path, model_cfg=model_cfg, device=device, allow_unsafe_pickle=allow_unsafe_pickle
-        )
-        return cls(module, device=device, task_kind=getattr(model_cfg, "task_kind", None))
-
-    @staticmethod
-    def _infer_task_kind(model: nn.Module) -> TaskKind:
-        from raitap.task_families import TASK_FAMILIES
-
-        matches = [
-            family.kind
-            for family in TASK_FAMILIES.values()
-            if getattr(family, "matches_model", None) and family.matches_model(model)
-        ]
-        if len(matches) > 1:
-            raise ValueError(
-                f"Model matches multiple task families {sorted(k.value for k in matches)}; "
-                "set model.task_kind explicitly to disambiguate."
-            )
-        return matches[0] if matches else TaskKind.classification
-
-    @property
-    def task_kind(self) -> TaskKind:
-        return self._task_kind
-
-    @property
-    def hardware_label(self) -> str:
-        return _torch_hardware_label(self.device)
-
-    def _prepare_inputs(self, inputs: torch.Tensor) -> torch.Tensor:
-        reshaped = _adapt_input_shape(inputs, self.expected_input_shape)
-        return reshaped.to(self.device)
-
-    def prepare_detection_inputs(self, inputs: list[torch.Tensor]) -> list[torch.Tensor]:
-        """Prepare a ragged list of detection images for inference.
-
-        Each ``(C, H, W)`` tensor is moved to the backend device without
-        any shape adaptation — detection images keep their native resolution.
-        Torchvision detection models accept ``list[Tensor]`` natively.
-        """
-        return [img.to(self.device) for img in inputs]
-
-    def _prepare_kwargs(self, kwargs: dict[str, Any]) -> dict[str, Any]:
-        return _move_tensors_to_device(kwargs, self.device)
-
-    def __call__(self, inputs: torch.Tensor | list[torch.Tensor]) -> Any:
-        return self.model(inputs)
-
-    def autograd_module(self) -> nn.Module:
-        return self.model
-
-
-@register(
-    provides=FORWARD_ONLY,
-    extensions={".onnx"},
-    extra="onnx",
-    supported_hardware={ResolvedHardware.cpu, ResolvedHardware.cuda, ResolvedHardware.xpu},
-)
-class OnnxBackend(ModelBackend):
-    """ONNX Runtime-backed model runtime."""
-
-    def __init__(
-        self,
-        session: ort.InferenceSession,
-        *,
-        providers: Sequence[str],
-        model_path: Path | None = None,
-    ) -> None:
-        self.session = session
-        self.providers = list(providers)
-        self.model_path = model_path
-        self._preprocessing_module: nn.Module | None = None
-        inputs = session.get_inputs()
-        if len(inputs) != 1:
-            raise ValueError(
-                f"Expected exactly one ONNX input, got {len(inputs)}. "
-                "Multi-input ONNX models are not supported yet."
-            )
-        self.input_name = inputs[0].name
-        self.input_type = inputs[0].type
-        self.output_names = [output.name for output in session.get_outputs()]
-        self.expected_input_shape = _resolve_onnx_expected_shape(inputs[0].shape)
-
-    @property
-    def hardware_label(self) -> str:
-        primary_provider = self.providers[0] if self.providers else "CPUExecutionProvider"
-        return _onnx_hardware_label(primary_provider)
-
-    def set_preprocessing(self, module: nn.Module | None) -> None:
-        """Attach a preprocessing module for tensor calls."""
-        if module is not None:
-            module.cpu()
-            module.eval()
-        self._preprocessing_module = module
-
-    def _prepare_inputs(self, inputs: torch.Tensor) -> torch.Tensor:
-        """Leave shape adaptation to attached preprocessing when present."""
-        if self._preprocessing_module is not None:
-            return inputs
-        return super()._prepare_inputs(inputs)
-
-    @classmethod
-    def from_path(
-        cls,
-        path: Path,
-        *,
-        model_cfg: Any = None,
-        hardware: str = "gpu",
-        allow_unsafe_pickle: bool = False,
-    ) -> OnnxBackend:
-        # model_cfg / allow_unsafe_pickle unused: ONNX needs neither.
-        try:
-            import onnxruntime as ort
-        except ImportError as error:
-            raise ImportError(_ONNX_RUNTIME_INSTALL_HINT) from error
-
-        providers = resolve_onnx_providers(
-            hardware,
-            available_providers=ort.get_available_providers(),
-        )
-        session = ort.InferenceSession(str(path), providers=providers)
-        return cls(session, providers=providers, model_path=path)
-
-    def forward_numpy(self, batch: np.ndarray[Any, Any]) -> np.ndarray[Any, Any]:
-        """Run inference on a NumPy batch (no Torch required on the hot path)."""
-        expected_dtype = _NUMPY_DTYPES_BY_ONNX_TYPE.get(self.input_type)
-        input_array = _adapt_input_shape(batch, self.expected_input_shape)
-        if expected_dtype is not None and input_array.dtype != expected_dtype:
-            input_array = input_array.astype(expected_dtype, copy=False)
-        output_names: list[str] | None = self.output_names or None
-        outputs = self.session.run(output_names, {self.input_name: input_array})
-        np_outputs = [np.asarray(output) for output in outputs]
-        if not np_outputs:
-            raise ValueError("ONNX Runtime returned no outputs.")
-        if len(np_outputs) == 1:
-            return np_outputs[0]
-        return self._select_primary_numpy_output(np_outputs)
-
-    def __call__(self, inputs: torch.Tensor) -> torch.Tensor:
-        if not isinstance(inputs, torch.Tensor):
-            raise TypeError(
-                f"OnnxBackend expected torch.Tensor inputs, got {type(inputs).__name__}."
-            )
-
-        prepared_inputs = inputs
-        if self._preprocessing_module is not None:
-            with torch.no_grad():
-                prepared_inputs = self._preprocessing_module(inputs.detach().cpu())
-            if not isinstance(prepared_inputs, torch.Tensor):
-                raise TypeError(
-                    "OnnxBackend preprocessing expected a torch.Tensor output, "
-                    f"got {type(prepared_inputs).__name__}."
-                )
-
-        input_array = self._tensor_to_numpy(prepared_inputs)
-        primary = self.forward_numpy(input_array)
-        return torch.from_numpy(np.asarray(primary))
-
-    def _tensor_to_numpy(self, inputs: torch.Tensor) -> np.ndarray[Any, Any]:
-        array = inputs.detach().cpu().numpy()
-        expected_dtype = _NUMPY_DTYPES_BY_ONNX_TYPE.get(self.input_type)
-        if expected_dtype is None:
-            return array
-        if array.dtype != expected_dtype:
-            return array.astype(expected_dtype, copy=False)
-        return array
-
-    @staticmethod
-    def _select_primary_output(outputs: list[torch.Tensor]) -> torch.Tensor:
-        for tensor in outputs:
-            if tensor.ndim >= 2:
-                return tensor
-        return max(outputs, key=lambda tensor: tensor.numel())
-
-    @staticmethod
-    def _select_primary_numpy_output(outputs: list[np.ndarray[Any, Any]]) -> np.ndarray[Any, Any]:
-        for arr in outputs:
-            if arr.ndim >= 2:
-                return arr
-        return max(outputs, key=lambda arr: arr.size)
-
-
-def _move_tensors_to_device(value: Any, device: torch.device) -> Any:
-    if isinstance(value, torch.Tensor):
-        return value.to(device)
-    if isinstance(value, list):
-        return [_move_tensors_to_device(item, device) for item in value]
-    if isinstance(value, tuple):
-        return tuple(_move_tensors_to_device(item, device) for item in value)
-    if isinstance(value, dict):
-        return {key: _move_tensors_to_device(item, device) for key, item in value.items()}
-    return value
-
-
-def _torch_hardware_label(device: torch.device) -> str:
-    label_by_type = {
-        "cpu": "CPU",
-        "cuda": "CUDA",
-        "xpu": "Intel XPU",
-        "mps": "Apple MPS",
-    }
-    return label_by_type.get(device.type, device.type.upper())
-
-
-def _onnx_hardware_label(provider: str) -> str:
-    label_by_provider = {
-        "CPUExecutionProvider": "CPU",
-        "CUDAExecutionProvider": "CUDA",
-        "OpenVINOExecutionProvider": "Intel OpenVINO",
-        "CoreMLExecutionProvider": "Apple CoreML",
-    }
-    return label_by_provider.get(provider, provider)
