+++ src/raitap/robustness/tests/test_factory.py
@@ -9,7 +9,7 @@ import torch
 from omegaconf import OmegaConf
 
 from raitap.configs.adapter_factory import resolve_call_data_sources
-from raitap.models.backend import ModelBackend
+from raitap.models.base_backend import ModelBackend
 from raitap.robustness.assessors import TorchattacksAssessor
 from raitap.robustness.contracts import AssessmentKind
 from raitap.robustness.exceptions import AssessmentKindVisualiserIncompatibilityError
